fix up repl
