+++ ruff.toml
@@ -106,7 +106,9 @@ lint.per-file-ignores."tests/**/*.py" = [
   "TID252",
 ]
 lint.per-file-ignores."tests/reqwest/conftest.py" = ["ARG001"]
-lint.per-file-ignores."python/ry/dev.py" = []
+lint.per-file-ignores."python/ry/dev/*.py" = [
+  "PLC0415"
+]
 lint.per-file-ignores."python/ry/ryo3/*.pyi" = ["F401", "F821"]
 lint.per-file-ignores."scripts/**" = ["S101", "S603", "T201"]
 lint.unfixable = []
